Upgrade HPS Wrapper Demo to Quartus v22.1
Update the Quartus project and IP set for the HPS wrapper to the latest versions.

The design of the wrapper and its software image has been internally changed to support the latest version of the bootloader used in the ELEC5620M teaching.

There are no user facing changes - the top level pinout and behaviour of the wrapper are unchanged.